chg: considerable change in the Numerator project: is now independent on chosen graphics unit
--HG--
extra : convert_revision : svn%3Ad1681fee-8e23-fc4c-b001-5eb8f207e95b/trunk%40446

+++ Libraries/Numerator/MathML.Rendering/GlyphMap.cs
@@ -183,7 +183,7 @@ namespace MathML.Rendering.GlyphMapper
 		 */
     private int GetFontIndex(IFormattingContext context, float fontSize)
 		{
-			fontSize = (int)(fontSize * scaleFactor);
+			fontSize = (fontSize * scaleFactor);
 			for(int i = 0; i < fontInstances.Length; i++)
 			{
 				if(fontInstances[i].PointSize == fontSize)
@@ -202,7 +202,7 @@ namespace MathML.Rendering.GlyphMapper
 		 */
     private int CreateFont(IFormattingContext context, float fontSize)
 		{
-			fontSize = (int)(fontSize * scaleFactor);
+			fontSize = (fontSize * scaleFactor);
 			FontInstance[] tmp = fontInstances;
 			fontInstances = new FontInstance[fontInstances.Length + 1];	
 			tmp.CopyTo(fontInstances, 0);
